requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coverage==5.3
cryptography==3.1.1
Django==3.1.1
django-cacheops==5.0.1
django-celery-beat==2.1.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-environ==0.4.5
django-redis==4.12.1
django-redis-sessions==0.6.1
django-timezone-field==4.0
djangorestframework==3.11.1
funcy==1.15
idna==2.10
jwt==1.0.0
kombu==5.0.2
mysqlclient==2.0.1
prompt-toolkit==3.0.7
psycopg2-binary==2.8.5
pycparser==2.20
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5

PyJWT~=1.7.1
djangorestframework-jwt~=1.11.0